mastro_pull/pyproject.toml
[tool.poetry]
name = "mastro-pull"
version = "0.1.0"
description = ""
authors = ["Alvaro Garcia Molino <alvaro.garcia.molino@acciona.com>"]
readme = "README.md"
packages = [{include = "mastro_pull"}]

[tool.poetry.dependencies]
python = "^3.11"
rich = "^13.7.1"
numpy = "^1.24.2"
pika = "^1.3.1"
psycopg2-binary = "^2.9.5"
pymodbus = "^3.2.0"
pyyaml = "^6.0"
pandas = "^2.1.0"
pydantic = "^2.5.2"
pydantic-settings = "^2.1.0"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.2"
ruff = "^0.1.9"
requests = "^2.31.0"
pendulum = "^3.0.0"
justpy = "^0.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
